Revert "ADB raw packets"
This reverts commit 1e3e85c3842a3aac50bed90bb9c3537944bbc40d.

From Finn Thain <fthain@telegraphics.com.au>:

Geert, you should drop this patch, because there are no other patches in
your queue that use ADBREQ_RAW and because it seems to be broken (the idea
of sending a raw packet using adb_request() was apparently to permit
req->data[0] != ADB_PACKET). Also it seems to have become obsolete back in
2006.

I went digging in the mac68k repo. ADBREQ_RAW shows up here:

    "Latest ADB changes including the new 68k PMU driver."
    http://linux-mac68k.cvs.sourceforge.net/viewvc/linux-mac68k/linux-mac68k/drivers/macintosh/adb.c?r1=1.15&r2=1.16&

    "Added Baboon interrupt support.
    Moved all time and PRAM-related functions into misc.c."
    http://linux-mac68k.cvs.sourceforge.net/viewvc/linux-mac68k/linux-mac68k/arch/m68k/mac/misc.c?r1=1.4&r2=1.5&

You are right, ADBREQ_RAW was clearly PRAM and clock related.

And this seems to indicate roughly when it got broken:

    "Merge with 2.4.0 release."
    http://linux-mac68k.cvs.sourceforge.net/viewvc/linux-mac68k/linux-mac68k/drivers/macintosh/adb.c?r1=1.21&r2=1.22

Anyway, in mainline kernels, adb_request() was not called from the PRAM
and clock related code since 2006 when Al Viro got that code to build
again:

    https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/arch/m68k/mac/misc.c?id=3272244c2b1a8f13cec83c04b8245fa7fcb47a27

From Brad Boyer <flar@allandria.com>:

Thank you. I took a look, and I agree with everyone else that this was a
legacy patch that isn't required.

The original idea was to use the same API for both real ADB requests and
everything else that goes through the PMU, Egret, or CUDA chips. This
includes not only RTC/PRAM management, but power control and several
other things as well. It wasn't such a great idea, honestly.

Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>